Consolidated IRace and IClass files
