+++ DndUtils/CharacterGenerator/CharacterModel.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace DndUtils
+namespace DndUtils.CharacterGenerator
 {
     class CharacterModel
     {
