+++ DndUtils/CharacterGenerator/CharacterView.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DndUtils
+namespace DndUtils.CharacterGenerator
 {
     class CharacterView
     {
