+++ DndUtils/CharacterGenerator/Race/Dragonborn.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DndUtils.Race
+namespace DndUtils.CharacterGenerator.Race
 {
     class Dragonborn : IRace
     {
@@ -20,5 +20,20 @@ namespace DndUtils.Race
         };
         protected bool BaseDragonbornDarkvision = false;
         protected HashSet<string> BaseDragonbornProficiencies = new HashSet<string>();
+    }
+
+    class BaseDragonborn : Dragonborn
+    {
+        public BaseDragonborn()
+        {
+            _raceName = "Dragonborn";
+            _raceScoreBuff = new Dictionary<string, int>(BaseDragonbornASI);
+            _raceSize = BaseDragonbornSize;
+            _raceSpeed = BaseDragonbornSpeed;
+            _raceLanguages = BaseDragonbornLanguages;
+            _darkvision = BaseDragonbornDarkvision;
+            _raceProficiencies = new HashSet<string>(BaseDragonbornProficiencies);
+            _sourceBook = "Player's Handbook";
+        }
     }
 }
