+++ DndUtils/CharacterGenerator/Race/HalfElf.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DndUtils.Race
+namespace DndUtils.CharacterGenerator.Race
 {
     class HalfElf : IRace
     {
@@ -19,5 +19,20 @@ namespace DndUtils.Race
         };
         protected bool BaseHalfElfDarkvision = true;
         protected HashSet<string> BaseHalfElfProficiencies = new HashSet<string>();
+    }
+
+    class BaseHalfElf : HalfElf
+    {
+        public BaseHalfElf()
+        {
+            _raceName = "Half-Elf";
+            _raceScoreBuff = new Dictionary<string, int>(BaseHalfElfASI);
+            _raceSize = BaseHalfElfSize;
+            _raceSpeed = BaseHalfElfSpeed;
+            _raceLanguages = BaseHalfElfLanguages;
+            _darkvision = BaseHalfElfDarkvision;
+            _raceProficiencies = new HashSet<string>(BaseHalfElfProficiencies);
+            _sourceBook = "Player's Handbook";
+        }
     }
 }
