+++ DndUtils/CharacterGenerator/Race/HalfOrc.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DndUtils.Race
+namespace DndUtils.CharacterGenerator.Race
 {
     class HalfOrc : IRace
     {
@@ -23,5 +23,20 @@ namespace DndUtils.Race
         {
             "Intimidation"
         };
+    }
+
+    class BaseHalfOrc : HalfOrc
+    {
+        public BaseHalfOrc()
+        {
+            _raceName = "Half-Orc";
+            _raceScoreBuff = new Dictionary<string, int>(BaseHalfOrcASI);
+            _raceSize = BaseHalfOrcSize;
+            _raceSpeed = BaseHalfOrcSpeed;
+            _raceLanguages = BaseHalfOrcLanguages;
+            _darkvision = BaseHalfOrcDarkvision;
+            _raceProficiencies = new HashSet<string>(BaseHalfOrcProficiencies);
+            _sourceBook = "Player's Handbook";
+        }
     }
 }
