+++ DndUtils/CharacterGenerator/Race/Tiefling.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DndUtils.Race
+namespace DndUtils.CharacterGenerator.Race
 {
     class Tiefling : IRace
     {
@@ -20,5 +20,20 @@ namespace DndUtils.Race
         };
         protected bool BaseTieflingDarkvision = true;
         protected HashSet<string> BaseTieflingProficiencies = new HashSet<string>();
+    }
+
+    class BaseTiefling : Tiefling
+    {
+        public BaseTiefling()
+        {
+            _raceName = "Tiefling";
+            _raceScoreBuff = new Dictionary<string, int>(BaseTieflingASI);
+            _raceSize = BaseTieflingSize;
+            _raceSpeed = BaseTieflingSpeed;
+            _raceLanguages = BaseTieflingLanguages;
+            _darkvision = BaseTieflingDarkvision;
+            _raceProficiencies = new HashSet<string>(BaseTieflingProficiencies);
+            _sourceBook = "Player's Handbook";
+        }
     }
 }
